CHET-452: [core] move aws cleanup implementation into new cleanup package
